Add trailers to MovieDetailsActivity
Added methods for using a Loaders to manage the syncing and Cursors for
the movie details and trailers.

Added TrailerAdapter for displaying trailers in the Activity.

Added animations for the favorite icon as the movie poster is scrolled
into the content layout.

Added AndroidTest for testing adding and querying the database for
trailers.