Update layout for review list item
Previous layout attempted to wrap an entire review to its full length,
however if the next review was longer, the RecyclerView would only show
it up to the length of the first review. Modified it to show a certain
number of lines of review content. The entire review will become visible
in a Dialog when the user clicks on it.