Fix movie sort order
Previously, the Dialog for changing sort methods would launch a new
AsyncTask to load the movie data. This has been deprecated in favor of
launching an IntentService to download the data in the background and
altering the sort method in SharedPreferences so that the sort method is
saved between user sessions.