Add save scroll state
Added a method to save the scroll state of the RecyclerView in
MovieListActivity. Needed a workaround due to the the wrap_content
property of the movie poster list item. wrap_content caused the
ViewHolder pattern height to be set to 0 upon its creation. When the
scroll state was restored, the scroll was restored to the position given
every element had a height of 0.

Workaround required setting a minHeight to the ViewHolder's layout and
then later on setting the correct height of each ViewHolder after the
correct height had been calculated.